New Backend Deployment - Deleted the old server directory to re-deploy the backend through Firebase. I was bad commiting during my process, but in this commit these api endpoints were deployed. The test endpoint worked instantly. The teacher's settings endpoint would only show an empty object but after abour of deployment it showed a specific teachers settings information.
