--- server/database-rules.json
@@ -1,6 +1,0 @@
-{
-  "rules": {
-    ".read": "auth != null",
-    ".write": "auth != null"
-  }
-}
