--- server/firestore.indexes.json
@@ -1,3 +1,0 @@
-{
-  "indexes": []
-}
