+++ server/functions/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "functions",
+  "description": "Cloud Functions for Firebase",
+  "scripts": {
+    "serve": "firebase serve --only functions",
+    "shell": "firebase functions:shell",
+    "start": "nodemon index.js",
+    "deploy": "firebase deploy --only functions",
+    "logs": "firebase functions:log"
+  },
+  "dependencies": {
+    "@google-cloud/firestore": "^0.19.0",
+    "cors": "^2.8.5",
+    "express": "^4.16.4",
+    "firebase-admin": "^6.2.0",
+    "firebase-functions": "^2.1.0"
+  },
+  "private": true,
+  "devDependencies": {
+    "nodemon": "^1.18.7"
+  }
+}
