--- server/index.html
@@ -1,13 +1,0 @@
-<html lang="en">
-<head>
-	<meta charset="utf-8">
-	<title>Home</title>
-	<link href="https://fonts.googleapis.com/css?family=Roboto:400,700" rel="stylesheet">
-	<link rel="stylesheet" href="css/index.css">
-	<!--[if IE]>
-		<script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-	<![endif]-->
-</head>
-
-  <body>Firebase backend deployed! This is just a confirmation page that it is running, see our <a href = 'https://musicmaker-teacher.netlify.com/'>front end</a> instead!</body>
-</html>
