--- server/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "MusicMaker",
-  "version": "1.0.0",
-  "description": "A web application for music teachers to view and grade student performance practice with an associated iOS app for student video submission.",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon server.js",
-    "qrcode": "nodemon qrcode.js"
-  },
-  "repository": "https://github.com/Lambda-School-Labs/Labs8-MusicMaker.git",
-  "author": "Keiran <keirankozlowski@gmail.com>",
-  "license": "MIT",
-  "dependencies": {
-    "@google-cloud/firestore": "^0.18.0",
-    "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
-    "cors": "^2.8.5",
-    "express": "^4.16.4",
-    "firebase": "^5.5.7",
-    "firebase-admin": "^5.12.1",
-    "firebase-functions": "^1.0.3",
-    "firebase-tools": "^6.0.1",
-    "firebaseui": "^3.4.1",
-    "qrcode": "^1.3.2",
-    "stripe": "^6.15.0"
-  },
-  "devDependencies": {
-    "nodemon": "^1.18.6"
-  }
-}
